Initial commit.
Wanted to try some material design so i changed the look and flow of the app.
Removed the navigation drawer, not a nice user experience in this app.
Navigation drawer is replaced with a FAB.

Used RecyclerView for the list of books and made a new adapter that the RecyclerView can use.

Changed how the adapter is reloaded, did a lot of unnecessary reloading.

Fixed so rotation handling works properly.

Cleared res for unused resources.

Refactoring and changes to layouts.

Added scanning functionality for book ISBN

Corrected the  ISBN-10 and 13 implementation, ISBN-10 was faulty.

 Updated icons.

 Handling image loading with Picasso instead of a task.

 Handle no connection when trying to fetch a book.

 Simple user message on errors.

 Fixed so settings work and default value is set on first launch.
 Setting is updated when set so selection can be seen by the user.

 Moved handling of large device to bool.xml file i n res instead.

 And probably something else that i forgot to write down...